FileDesc module changed to 'Sys'
- AlignedStorage module has been deleted, with code moved into the
  'Traits' module instead.
- FileDesc module has been renamed, to 'Sys'. This will be more
  appropriate, as further code that belongs together with FileDesc is
  planned:
        - File (a wrapper around a FileDesc that refers specifically to
          a file)
        - Path (A wrapper around a Str that makes explicit that a Path
        is expected)
        - BufReader (Wraps around a 'Reader' (such as Socket or File),
          to allow reading into a buffer to optimise read operations)
