--- src/AlignedStorage/src/CMakeLists.txt
@@ -1,7 +1,0 @@
-# AlignedStorage is header only
-
-# file(GLOB_RECURSE SOURCES LIST_DIRECTORIES true *.hpp *.cpp)
-#
-# set(SOURCES ${SOURCES})
-#
-# add_library(${MODULE_NAME}_lib STATIC ${SOURCES})
