--- src/AlignedStorage/tst/main.cpp
@@ -1,6 +1,0 @@
-#include "gtest/gtest.h"
-
-int main(int argc, char** argv) {
-    ::testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
-}
