+++ src/CMakeLists.txt
@@ -4,24 +4,22 @@ set(SOURCES main.cpp)
 
 add_subdirectory(Utils)
 add_subdirectory(Str)
-add_subdirectory(FileDesc)
+add_subdirectory(Sys)
 add_subdirectory(Net)
 add_subdirectory(Option)
 add_subdirectory(Result)
 add_subdirectory(Traits)
-add_subdirectory(AlignedStorage)
 
 # Only for testing
 add_subdirectory(Examples)
 
 include_directories(Utils)
 include_directories(Str)
-include_directories(FileDesc)
+include_directories(Sys)
 include_directories(Net)
 include_directories(Option)
 include_directories(Result)
 include_directories(Traits)
-include_directories(AlignedStorage)
 
 add_executable(${BINARY}.run ${SOURCES})
