--- src/FileDesc/CMakeLists.txt
@@ -1,6 +1,0 @@
-set(MODULE_NAME FileDesc)
-
-include_directories(src)
-
-add_subdirectory(src)
-add_subdirectory(tst)
