+++ src/Net/src/CMakeLists.txt
@@ -7,5 +7,5 @@ add_library(${MODULE_NAME}_lib STATIC ${SOURCES})
 target_link_libraries(${MODULE_NAME}_lib
     PUBLIC Str_lib
     PUBLIC Utils_lib
-    PUBLIC FileDesc_lib
+    PUBLIC Sys_lib
     )
